added PROGMEM keyword, added braced init.
PROGMEM is used to denote a variable should be stored in flash memory
instead of runtime ram, no change was reported after the update. It is
likely that compilation will consider static constexpr to mean the same
thing, but it's a good idea to use where available. Braced init will
cause compilation to report where a constant value is being truncated to
fit into a smaller data type.